moved game_screen to new cpp
made header
