+++ fifteen_game.h
@@ -1,5 +1,4 @@
 #include "Tile_button.h"
 #include "Project_window.h"
 #include "End_screen.h"
-#include "Player_score.h"
-//note : move this include to game_screen later
+#include "Game_screen.h"
